[MOD] Replaced CardDeckInfoTable with RetentionTable
For the review feature, it is necessary to retrieve the cards due for the day.
Previously, retention data was stored as a JSON string in a TEXT column of the CardDeckInfoTable, which made it difficult to query without duplicating relevant fields at the root level.

The following changes were made:
- Removed CardDeckInfoTable
- Moved the deck ID to the CardsTable
- Stored retention information in a new RetentionTable